requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
tensorflow==2.12.0
xgboost==1.7.5
lightgbm==3.3.5

# Django backend
django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.0.0
drf-yasg==1.21.5
psycopg2-binary==2.9.6

# FastAPI backend
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
sqlalchemy==2.0.15

# Data pipeline
selenium==4.9.1
beautifulsoup4==4.12.2
requests==2.30.0
webdriver-manager==3.8.6

# LLM integration
ollama==0.1.2
langchain==0.0.200

# Testing
pytest==7.3.1
pytest-django==4.5.2

# Deployment
docker==6.1.1
gunicorn==20.1.0
